refactor: rm

+++ src/handlers/rm.js
@@ -1,11 +1,13 @@
 import { unlink } from 'fs/promises';
+import { existsSync } from 'fs';
 
 export const rm = async (input) => {
   try {
-    const filePath = input.split(' ')[1];
-    await unlink(filePath)
+    const filePath = input.slice(3).trim();
+    if (!existsSync(filePath)) throw new Error('Invalid input: no such file or directory');
+    await unlink(filePath);
     console.log(`File ${filePath} has been deleted.`);
   } catch (error) {
-    console.error(error);
+    console.error(error.message);
   }
 };
